--- a/wolfcrypt/src/aes.c
+++ b/wolfcrypt/src/aes.c
@@ -1691,6 +1691,13 @@ static void wc_AesEncrypt(Aes* aes, const byte* inBlock, byte* outBlock)
     return;
 #endif
 
+#if defined(WOLFSSL_IMXRT_DCP)
+    if (aes->keylen == 16) {
+        DCPAesEcbEncrypt(aes, outBlock, inBlock, AES_BLOCK_SIZE);
+        return;
+    }
+#endif
+
     /*
      * map byte array block to cipher state
      * and add initial round key:
@@ -1972,6 +1979,12 @@ static void wc_AesDecrypt(Aes* aes, const byte* inBlock, byte* outBlock)
 #endif /* WOLFSSL_AESNI */
 #if defined(WOLFSSL_SCE) && !defined(WOLFSSL_SCE_NO_AES)
     return AES_ECB_decrypt(aes, inBlock, outBlock, AES_BLOCK_SIZE);
+#endif
+#if defined(WOLFSSL_IMXRT_DCP)
+    if (aes->keylen == 16) {
+        DCPAesEcbDecrypt(aes, outBlock, inBlock, AES_BLOCK_SIZE);
+        return;
+    }
 #endif
 
     /*
@@ -2531,7 +2544,6 @@ static void wc_AesDecrypt(Aes* aes, const byte* inBlock, byte* outBlock)
 #elif defined(WOLFSSL_DEVCRYPTO_AES)
     /* implemented in wolfcrypt/src/port/devcrypto/devcrypto_aes.c */
 
-
 #else
 
     /* Software AES - SetKey */
@@ -2853,6 +2865,13 @@ static void wc_AesDecrypt(Aes* aes, const byte* inBlock, byte* outBlock)
 
         #ifdef WOLFSSL_IMX6_CAAM_BLOB
             ForceZero(local, sizeof(local));
+        #endif
+        #ifdef WOLFSSL_IMXRT_DCP
+            ret = 0;
+            if (keylen == 16)
+                ret = DCPAesSetKey(aes, userKey, keylen, iv, dir);
+            if (ret != 0)
+                return WC_HW_E;
         #endif
 
             return ret;
@@ -4198,6 +4217,14 @@ int wc_AesGcmSetKey(Aes* aes, const byte* key, word32 len)
 
 #ifdef WOLFSSL_IMX6_CAAM_BLOB
     ForceZero(local, sizeof(local));
+#endif
+
+#ifdef WOLFSSL_IMXRT_DCP
+    ret = 0;
+    if (len == 16)
+        ret = DCPAesSetKey(aes, key, len, iv, AES_ENCRYPTION);
+    if (ret != 0)
+        return WC_HW_E;
 #endif
 
     return ret;