@josepho0918 2020-09-21
A few concerns for dcp implementation:
1. Mutex protection might be required for DCP hardware.
2. "s_dcpContextSwitchingBuffer" is a shared buffer in fsl_dcp.c. It makes hash functions not reentrant.
3. DCP hash doesn't work properly if ctx buffer is not 32-byte aligned when using cache.