@danielinux 2020-09-22
@josepho0918 I've updated the driver:

1. DCP is now protected by a mutex when running on multi-threaded environments
2. Context switching allows for 4 parallel channel configuration to be saved/restored. This driver is now using the 4 channels available and keeping track of them upon init/free. A `WC_PENDING_E` is generated when there are no channels available.
3. DCACHE mode is not supported by this version of the driver. An error is generated if the code is compiled with `DCP_USE_DCACHE`.